@phosphor-icons/react to dev dependencies
