+++ ai-chatbot/package.json
@@ -10,7 +10,7 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@likeminds.community/likeminds-chat-reactjs": "1.3.0",
+    "@likeminds.community/chat-js-beta": "^1.1.197",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-hot-toast": "^2.4.1",
