+++ ai-chatbot/src/App.tsx
@@ -6,6 +6,7 @@ import {
 } from "@likeminds.community/likeminds-chat-reactjs";
 import React from "react";
 import "./App.css";
+import LMChatClient from "@likeminds.community/chat-js-beta";
 
 const App = () => {
   const [userDetails, setUserDetails] = useState<{
@@ -17,8 +18,14 @@ const App = () => {
     apiKey?: string;
   }>({});
 
-  const lmChatClient = initiateLMClient();
+  // const lmChatClient = initiateLMClient();
+  const VERSION_CODE = 42;
+  const PLATFORM_CODE = "rt";
 
+  const lmChatClient = LMChatClient.setPlatformCode(PLATFORM_CODE)
+    .setVersionCode(VERSION_CODE)
+    .setExcludedConversationStates([])
+    .build();
   useEffect(() => {
     const queryString = window.location.search;
     const params = new URLSearchParams(queryString);
@@ -35,7 +42,13 @@ const App = () => {
   return (
     <LMChatAIButton
       client={lmChatClient}
-      userDetails={userDetails}
+      userDetails={{
+        apiKey: "aa2a3a49-f371-45de-a071-7cafc1fa927a",
+        uuid: "NEW_ANUJ_USER_04",
+        username: "NEW_ANUJ_USER_04",
+      }}
+      lmChatCoreCallbacks={}
+      // userDetails={userDetails}
       lmChatTheme={LMChatTheme.NETWORKING_CHAT}
     />
   );
