+++ community-chat/src/App.tsx
@@ -3,11 +3,8 @@ import {
   LMClientOverlayProvider,
   LMChannel,
   initiateLMClient,
-  LMChatCurrentMode,
   LMChatTheme,
 } from "@likeminds.community/likeminds-chat-reactjs";
-import LMChatClient from "@likeminds.community/chat-js-beta";
-
 const App = () => {
   const [userDetails, setUserDetails] = useState<{
     accessToken?: string;
@@ -18,13 +15,7 @@ const App = () => {
     apiKey?: string;
   }>({});
 
-  const VERSION_CODE = 42;
-  const PLATFORM_CODE = "rt";
-  const EXCLUDED_CONVERSATION_STATES = [] as number[];
-  const lmChatClient = LMChatClient.setPlatformCode(PLATFORM_CODE)
-    .setVersionCode(VERSION_CODE)
-    .setExcludedConversationStates(EXCLUDED_CONVERSATION_STATES)
-    .build();
+  const lmChatClient = initiateLMClient();
   useEffect(() => {
     const queryString = window.location.search;
     const params = new URLSearchParams(queryString);
