+++ community-hybrid-chat/src/main.tsx
